libs/agent_sdk/pyproject.toml
[project]
name = "agent-sdk"
version = "0.1.0"
description = "Shared infrastructure SDK for autonomous sentiment-analyst agents"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0",
    "pydantic>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",                     # OllamaProvider HTTP client
    "sqlalchemy[asyncio]>=2.0",        # IAsyncSession / ISessionFactory protocol
    "opensearch-py[async]>=2.0",       # OpenSearchAuditLogger
    "opentelemetry-sdk>=1.25",         # OllamaProvider tracing spans
    "opentelemetry-api>=1.25",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0",                      # RedisStreamQueue
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "sqlalchemy-stubs>=0.4",
]

# ---------------------------------------------------------------------------
# Build system
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_sdk"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
